# cuscuta-entry/src/main.rs
//! cuscuta 的 entry。
//!
//! # 依赖环境变量
//! cuscuta-entry 使用环境变量注入参数，这个 crate 依赖的环境变量有：
//! - `REDIS_STREAM_REFRESH_TTL`
//! - `REDIS_ADDR`
//! - `ACCOUNTS_SQL_ADDR`
//! - `STAT_ENABLE`
//! - `RESOURCE_UPDATE_PERIOD`（数据刷新周期，单位秒，默认 30）
//! - `RESOURCE_UPDATE_RETRIES`（数据拉取最大重试次数，默认 5）
//! - `RESOURCES_SONG_URL`

mod data;
mod db;
mod endpoints;
mod init;
mod loop_tasks;

use crate::{
    data::{CONFIG, SONG_LIST},
    db::{postgresql::POSTGRESQL_POOL, redis::REDIS_CLIENT},
    endpoints::{query::query, status::stat},
    enqueue::enqueue,
    loop_tasks::sync_config,
};
use cuscuta_common::{
    data::read_parsed_env,
    quick_fetch::QuickFetch,
    scheduled_job::{
        register_job_future,
        tasks::{open_postgresql_client, open_redis_client, sync_song_list},
    },
};

use axum::{
    Json, Router,
    response::IntoResponse,
    routing::{get, post},
};
use cuscuta_common::{
    batch_check_initialized,
    scheduled_job::{register_individual_job, register_job},
};
use reqwest::StatusCode;
use serde_json::json;
use tokio::net::TcpListener;
use tokio_util::sync::CancellationToken;
use tower_http::trace::{self, TraceLayer};
use tracing::Level;

use crate::{endpoints::enqueue, init::cuscuta_init};

#[tokio::main]
async fn main() {
    tracing_subscriber::fmt()
        .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
        .init();
    tracing::info!("starting...");
    let halt_token = CancellationToken::new();
    let trace_layer = TraceLayer::new_for_http()
        .make_span_with(trace::DefaultMakeSpan::new().level(Level::INFO))
        .on_request(trace::DefaultOnRequest::new().level(Level::INFO))
        .on_response(trace::DefaultOnResponse::new().level(Level::INFO));
    let service = Router::new()
        .route("/healthz", get(healthz))
        .route("/readyz", get(readyz))
        .route("/v1/enqueue", post(enqueue).layer(trace_layer.clone()))
        .route("/v1/query", get(query).layer(trace_layer.clone()))
        .route("/v1/status", get(stat));
    let addr = TcpListener::bind("0.0.0.0:8081")
        .await
        .expect("failed to bind 0.0.0.0:8081");
    tracing::info!("listening in 0.0.0.0:8081...");
    let data_update_period = read_parsed_env::<u64>("RESOURCE_UPDATE_PERIOD").unwrap_or_else(|e| {
        tracing::info!(
            "pre_init: failed to read RESOURCE_UPDATE_PERIOD: {e}, set to default (30s)"
        );
        30
    });
    tokio::spawn(register_individual_job(
        halt_token.clone(),
        CancellationToken::new(),
        10,
        cuscuta_init,
    ));
    tokio::spawn(register_job_future(
        halt_token.clone(),
        10,
        open_redis_client(&REDIS_CLIENT),
    ));
    tokio::spawn(register_job_future(
        halt_token.clone(),
        10,
        open_postgresql_client(&POSTGRESQL_POOL),
    ));
    tokio::spawn(register_job_future(
        halt_token.clone(),
        data_update_period,
        sync_song_list(&SONG_LIST),
    ));
    tokio::spawn(register_job(halt_token.clone(), 10, sync_config));
    axum::serve(addr, service)
        .with_graceful_shutdown(shutdown_signal(halt_token))
        .await
        .unwrap_or_else(|e| panic!("{e:?}"));
}

fn check_ready() -> Option<&'static str> {
    if REDIS_CLIENT.get().is_none() {
        return Some("redis client is not initialized");
    }
    batch_check_initialized!(CONFIG, SONG_LIST, POSTGRESQL_POOL);
    None
}

async fn readyz() -> impl IntoResponse {
    check_ready().map_or_else(
        || (StatusCode::OK, Json(json!({"status":"ready"}))),
        |e| {
            (
                StatusCode::SERVICE_UNAVAILABLE,
                Json(json!({"status": "not ready", "reason": e})),
            )
        },
    )
}

async fn healthz() -> impl IntoResponse {
    (StatusCode::OK, Json(json!({"health":"ok"})))
}

#[allow(clippy::ignored_unit_patterns)]
async fn shutdown_signal(cancellation_token: CancellationToken) {
    #[cfg(target_os = "linux")]
    {
        use tokio::signal::{
            self,
            unix::{SignalKind, signal},
        };
        let mut sigterm =
            signal(SignalKind::terminate()).expect("failed to install SIGTERM handler");
        tokio::select! {
            _ = signal::ctrl_c() => {},
            _ = sigterm.recv() => {},
            _ = cancellation_token.cancelled() => {},
        }
    }
    #[cfg(target_os = "windows")]
    {
        use tokio::signal;
        tracing::info!("咱其实挺想知道谁会在Windows上跑这个的……");
        tokio::select! {
            _ = signal::ctrl_c() => {},
            _ = cancellation_token.cancelled() => {},
        }
    }
    tracing::error!("cuscuta-entry halting...");
    cancellation_token.cancel();
}