# docs/openapi.yaml
openapi: 3.0.1
info:
  title: cuscuta
  description: 'cuscuta-entry 的 API 文档'
  version: 0.1.14
paths:
  /v1/enqueue:
    post:
      summary: 任务入列
      description: 输入好友码，将任务加入队列
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                friend_code:
                  type: string
                  description: 好友码（9 位数字）
                  example: '123456789'
                info_only:
                  type: boolean
                  description: 是否仅获取玩家信息
                  example: false
              required:
                - friend_code
                - info_only
      responses:
        '200':
          description: 入列成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  query_token:
                    type: string
                    description: 用于查询的 token
                required:
                  - success
                  - query_token
              example:
                success: true
                query_token: MTIzNDU2Nzg5LTE3NjAwMDAwMDA=
        '400':
          description: 好友码格式错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                code: -502
                message: 'bad request:-502'
        '418':
          description: 没有可用的 worker（无活跃账号）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                code: -400
                message: 'service unavailable: 0x000:-400'
        '500':
          description: 服务未就绪或数据库/Redis 错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                code: -101
                message: 'service unavailable: 0x020:-101'
  /v1/query:
    get:
      summary: 查询任务
      description: 使用 enqueue 返回的 query_token 查询任务完成情况
      parameters:
        - name: token
          in: query
          required: true
          description: 查询 token（即 enqueue 返回的 query_token）
          schema:
            type: string
      responses:
        '200':
          description: 任务已完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pending:
                    type: boolean
                  friend_info:
                    allOf:
                      - $ref: '#/components/schemas/FriendInfo'
                    nullable: true
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/SongScore'
                required:
                  - success
                  - pending
                  - friend_info
                  - result
        '206':
          description: 任务仍在进行
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pending:
                    type: boolean
                  total_jobs:
                    type: integer
                    description: 任务总分片数
                  finished_jobs:
                    type: integer
                    description: 已完成的分片数
                  pending_jobs:
                    type: integer
                    description: 已被 worker 消费、正在执行的分片数
                  queueing_jobs:
                    type: integer
                    description: 仍在队列等待的分片数
                  friend_info:
                    allOf:
                      - $ref: '#/components/schemas/FriendInfo'
                    nullable: true
                  eta:
                    type: number
                    description: 估算剩余时间（当前实现固定返回 0）
                    nullable: true
                required:
                  - success
                  - pending
                  - total_jobs
                  - finished_jobs
                  - pending_jobs
                  - queueing_jobs
                  - friend_info
                  - eta
        '400':
          description: token 非法或校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 任务失败或服务错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/status:
    get:
      summary: 查询 worker 状态
      description: 查询集群中 worker 的工作状态
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  worker_status:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkerStatus'
                required:
                  - success
                  - worker_status
        '400':
          description: status 接口未启用（STAT_ENABLE 为 false）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                code: -503
                message: 'bad request:-503'
        '500':
          description: 服务未就绪或 Redis 错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      title: 失败响应
      properties:
        success:
          type: boolean
        code:
          type: integer
          format: int64
        message:
          type: string
      required:
        - success
        - code
        - message
    SongScore:
      type: object
      title: 分数信息
      properties:
        song_id:
          type: string
        user_id:
          type: integer
          format: int64
        difficulty:
          type: integer
          format: int64
        score:
          type: integer
          format: int64
        score_below_max:
          type: integer
          format: int64
        shiny_perfect_count:
          type: integer
          format: int64
        perfect_count:
          type: integer
          format: int64
        near_count:
          type: integer
          format: int64
        miss_count:
          type: integer
          format: int64
        clear_type:
          type: integer
          format: int64
        best_clear_type:
          type: integer
          format: int64
        health:
          type: integer
          format: int64
        time_played:
          type: integer
          format: int64
        name:
          type: string
          description: 玩家名
    FriendInfo:
      type: object
      title: 好友信息
      properties:
        name:
          type: string
        user_id:
          type: integer
          format: int64
        rating:
          type: integer
          format: int64
          description: 实际显示评级为 rating/10.0
        character:
          type: integer
          format: int64
        is_char_uncapped:
          type: boolean
        is_char_uncapped_override:
          type: boolean
      required:
        - name
        - user_id
        - rating
        - character
        - is_char_uncapped
        - is_char_uncapped_override
    WorkerStatus:
      type: object
      title: worker 状态
      properties:
        last_active_timestamp:
          type: integer
          format: int64
          description: 最后一次活跃的时间戳（毫秒）
        cursor:
          type: integer
          description: 当前工作指针
        sub_queue:
          allOf:
            - $ref: '#/components/schemas/SubQueue'
          nullable: true
          description: 当前工作分片
        jobs:
          type: array
          description: 当前任务
          items:
            $ref: '#/components/schemas/Job'
      required:
        - last_active_timestamp
        - cursor
        - sub_queue
        - jobs
    SubQueue:
      type: object
      title: 任务分片
      properties:
        name:
          type: string
        hash:
          type: string
        timestamp:
          type: integer
          format: int64
        segment:
          type: object
          description: 该分片占有的分块（from 到 to）
          properties:
            start:
              type: integer
            end:
              type: integer
          required:
            - start
            - end
      required:
        - name
        - hash
        - timestamp
        - segment
    Job:
      type: object
      title: 任务
      properties:
        job_id:
          type: string
          description: 任务在 Redis 队列中的 id
        essential:
          $ref: '#/components/schemas/JobEssential'
        sub_queue:
          $ref: '#/components/schemas/SubQueue'
        state:
          $ref: '#/components/schemas/JobState'
      required:
        - job_id
        - essential
        - sub_queue
        - state
    JobEssential:
      type: object
      title: 任务关键信息
      properties:
        friend_code:
          type: string
          description: 查询对象的好友码
        timestamp:
          type: string
          description: 任务入列时的时间戳
        cursor_start:
          type: integer
          format: int32
        cursor_length:
          type: integer
          format: int32
        retry_count:
          type: integer
          format: int32
        job_uid:
          type: string
          description: 任务的唯一 ID
      required:
        - friend_code
        - timestamp
        - cursor_start
        - cursor_length
        - retry_count
        - job_uid
    JobState:
      title: 任务状态
      description: 任务内部状态（序列化为带标签的枚举，Cleaned 直接序列化为字符串）
      oneOf:
        - type: string
          enum:
            - Cleaned
        - type: object
          properties:
            Pulled:
              type: object
              properties:
                start_timestamp:
                  type: integer
                  format: int64
              required:
                - start_timestamp
          required:
            - Pulled
        - type: object
          properties:
            Pending:
              type: object
              properties:
                friend_info:
                  $ref: '#/components/schemas/FriendInfo'
                current_length:
                  type: integer
                start_timestamp:
                  type: integer
                  format: int64
              required:
                - friend_info
                - current_length
                - start_timestamp
          required:
            - Pending
        - type: object
          properties:
            Finished:
              type: object
              properties:
                friend_info:
                  $ref: '#/components/schemas/FriendInfo'
                start_timestamp:
                  type: integer
                  format: int64
              required:
                - friend_info
                - start_timestamp
          required:
            - Finished
        - type: object
          properties:
            Failed:
              type: object
              properties:
                friend_info:
                  allOf:
                    - $ref: '#/components/schemas/FriendInfo'
                  nullable: true
                start_timestamp:
                  type: integer
                  format: int64
                failure_info:
                  $ref: '#/components/schemas/JobFailure'
              required:
                - friend_info
                - start_timestamp
                - failure_info
          required:
            - Failed
    JobFailure:
      type: object
      title: 任务失败信息
      properties:
        fail_type:
          $ref: '#/components/schemas/JobFailureType'
        resume_strategy:
          $ref: '#/components/schemas/JobFailureResuming'
        timestamp_millis:
          type: integer
          format: int64
      required:
        - fail_type
        - resume_strategy
        - timestamp_millis
    JobFailureResuming:
      type: string
      title: 任务失败恢复策略
      enum:
        - Drop
        - NoOp
    JobFailureType:
      title: 任务失败类型
      description: 失败原因（单元变体直接序列化为字符串，带参数变体序列化为对象）
      oneOf:
        - type: string
          enum:
            - FriendNotFound
            - Reenqueued
        - type: object
          properties:
            TargetKeyNotFound:
              type: string
          required:
            - TargetKeyNotFound
        - type: object
          properties:
            ApiError:
              type: string
          required:
            - ApiError
        - type: object
          properties:
            XxxxxxApiError:
              type: array
              description: 第一个元素为 HTTP 状态码，第二个元素为可选的 API 错误码
              items:
                type: integer
                format: int64
              minItems: 2
              maxItems: 2
          required:
            - XxxxxxApiError